chore: rename file-system module
